HARMONY-1363
Patch didn't apply cleanly due to changes since patch was cut.

Had to be modified to support launcher, and now we have a more
complete and predicatable mechanism for finding artifacts via
harmonyvm.properties.

Known issues:

1) Stack overflow exception stuff is broken.  I had to remove the assert
   in signals_ia32.cpp line 336.  Rana knows and will look.  I also
   disabled the StackTest.

2) Test init/ErrorInitInitiialze ran too slowly for my limited patience
   under debug build, so that's disabled as well.

3) LIke before, jars dont' work.  That's next.

4) Tomcat wont' run either.

5) the c-based unit tests won't even run for some reason - it's some kind
   of ant or config problem, but I really have trouble grokking our test
   framework anyway, so given we pass 99% of the smoke and new and cool kernel
   tests, I'm happy to check this in and focus on the problems
   collectively

6) Also had to apply HARMONY-1459 to fix the problem with -fPIC, our asm
   code and gcc 3.4.  Thank you Gregory for saving me lots of time
   figuring out that one.




git-svn-id: https://svn.apache.org/repos/asf/incubator/harmony/enhanced/drlvm/trunk@443504 13f79535-47bb-0310-9956-ffa450edef68

+++ vm/jitrino/src/shared/Counter.h
@@ -23,19 +23,22 @@
 #ifndef _COUNTER_H_
 #define _COUNTER_H_
 
+
+#include <stddef.h>
+
+
 namespace Jitrino 
 {
 
-class CountWriter
-{
+class CountWriter {
 public:
 
-	virtual ~CountWriter ()		{}
+    virtual ~CountWriter ()     {}
 
-	virtual void write (const char* key, const char*  value)    	= 0;
-	virtual void write (const char* key, int          value)		= 0;
-	virtual void write (const char* key, unsigned int value)		= 0;
-	virtual void write (const char* key, double       value)		= 0;
+    virtual void write (const char* key, const char*  value)        = 0;
+    virtual void write (const char* key, int          value)        = 0;
+    virtual void write (const char* key, size_t       value)        = 0;
+    virtual void write (const char* key, double       value)        = 0;
 };
 
 
@@ -43,19 +46,19 @@ class CounterBase
 {
 public:
 
-	CounterBase (const char* s = 0);
-	virtual ~CounterBase ();
+    CounterBase (const char* s = 0);
+    virtual ~CounterBase ();
 
-	void setName (const char* s)					{key = s;}
-	void link ();
-	virtual void write  (CountWriter&)				= 0;
+    void setName (const char* s)                    {key = s;}
+    void link ();
+    virtual void write  (CountWriter&)              = 0;
 
-	const char* key;
+    const char* key;
 
 //protected:
 
-	CounterBase* next;
-	static CounterBase* head;
+    CounterBase* next;
+    static CounterBase* head;
 };
 
 
@@ -64,16 +67,15 @@ class Counter : public CounterBase
 {
 public:
 
-	Counter (const char* s)							: CounterBase(s) {}
-	Counter (const char* s, const T& v)				: CounterBase(s), value(v) {}
+    Counter (const char* s)                         : CounterBase(s) {}
+    Counter (const char* s, const T& v)             : CounterBase(s), value(v) {}
 
-	/*virtual*/void write  (CountWriter& logs)		{logs.write(key, value);}
+    /*virtual*/void write  (CountWriter& logs)      {logs.write(key, value);}
 
-	operator T& ()									{return value;}
+    operator T& ()                                  {return value;}
 
-	T value;
+    T value;
 };
-
 
 
 } //namespace Jitrino
